shAn

+++ src/main/java/com/project/tobe/service/OrderService.java
@@ -1,7 +1,6 @@
 package com.project.tobe.service;
 
 import com.project.tobe.dto.*;
-import org.apache.ibatis.annotations.Param;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,7 +8,7 @@ import java.util.List;
 public interface OrderService {
 
     List<OrderHDTO> getOrder(OrderSearchDTO criteria); //jsy 조건조회
-    List<PriceDTO> getPrice(Integer iocn, String delDate); //jsy 주문등록 판매가 가져오기
+    List<PriceDTO> getPrice(Long iocn, String delDate); //jsy 주문등록 판매가 가져오기
     Long registOrder(OrderRegistDTO orderRegistDTO); //jsy 주문 등록하기
     String getMyName(String myId); //이름 얻기
